fix: multiply percentages by 100 internally to match FormatJS (#44)
If the parameter is a percent-style number, then we multiply the value
by 100. This is in keeping with the ECMA-402 draft, which specifies the
`Intl.NumberFormat` rules. When using `Intl.NumberFormat` to format
percentages, the number must first be multiplied by 100 before any
formatting occurs. See section 15.1.6 of ECMA-402, specifically step
5.b.[^1]

ECMA-402, however, doesn't define an API for MessageFormat, so FormatJS
implements this on their own, using `Intl.NumberFormat` to process any
number parameters it encounters.[^2] As a result, all number parameters
in ICU message syntax that specify the `::percent` stem (i.e., "{0,
number, ::percent}") have their values first multiplied by 100 before
formatting them.

This may not be considered a bug in FormatJS, since it is adhering to
the ECMA-402 specification. However, it does not follow the rules for
percentages as programmed in icu4c (the underlying library PHP uses),
so in order to match the expected output of FormatJS, we multiply
percent values by 100 before formatting them.

Oddly enough, PHP's `NumberFormatter` has the same rules, and it uses
the underlying ICU implementation of the number formatter:

    $nf = new NumberFormatter('en-US', NumberFormatter::PERCENT);
    echo $nf->format(25); // Produces "2,500%"

While:

    $mf = new MessageFormatter('en-US', '{0, number, ::percent}');
    echo $mf->format([25]); // Produces "25%"

So, one could argue this is a bug in the ICU implementation of the
percent number skeleton.

[^1]: https://tc39.es/ecma402/#sec-partitionnumberpattern
[^2]: https://formatjs.io/docs/core-concepts/icu-syntax/#number-type