chore: use MIT license for source code (#58)
Revert "chore: relicense source code under Apache-2.0 (#49)"

This reverts commit 3178901045bee99513bac15cbe2f75295da6cdfd.

I received some feedback from peers in the community that the use of
Apache License 2.0 is not friendly to projects that are licensed under
GPLv2. While Apache License 2.0 is compatible with GPLv3, it is not
compatible with GPLv2:

> Please note that this license is not compatible with GPL version 2,
> because it has some requirements that are not in that GPL version.
> These include certain patent termination and indemnification
> provisions.
>
> — https://www.gnu.org/licenses/license-list.html#apache2

Since several major PHP projects use GPLv2 (Drupal, WordPress, TYPO3,
etc.), it's recommended that we use a license that is friendlier to
these projects, so that they may choose to use this library, if desired.

The reason projects licensed under GPLv2 cannot "link against" projects
licensed under Apache License 2.0 is because the Apache license adds
restrictions on uses of the software that GPLv2 does not have. The
general rule of thumb for GPL use is: if a downstream project has a
license that imposes additional restrictions on the use of the software,
then it's not compatible, since the user might not be aware or able to
make the decision to use that software, due to the restrictions it
imposes.

GPLv3 does not present a problem, since it imposes the same restrictions
as Apache, and therefore, linking against software that uses Apache
License 2.0 does not impose additional restrictions on the user.

However, we can't make these big projects switch from GPLv2 to GPLv3, so
if we wish for the widest possible audience, we must change to a more
permissive license.

This change reverts back to use of MIT License (i.e., "Expat License").